+++ packages/0x.js/package.json
@@ -20,14 +20,14 @@
         "build:all": "run-p build:umd:prod build:commonjs; exit 0;",
         "pre_build": "run-s generate_contract_wrappers copy_artifacts",
         "copy_artifacts": "copyfiles -u 2 './src/compact_artifacts/**/*.json' ./lib/src/compact_artifacts",
-        "generate_contract_wrappers": "abi-gen --abis 'src/compact_artifacts/@(Exchange|Token|TokenTransferProxy|EtherToken|TokenRegistry|DummyToken).json' --template ../contract_templates/contract.handlebars --partials '../contract_templates/partials/**/*.handlebars' --output src/contract_wrappers/generated --backend ethers && prettier --write 'src/contract_wrappers/generated/**.ts'",
+        "generate_contract_wrappers": "abi-gen --abis 'src/compact_artifacts/@(Exchange|Token|TokenTransferProxy|EtherToken|TokenRegistry|DummyToken).json' --template ../contract_templates/contract.handlebars --partials '../contract_templates/partials/**/*.handlebars' --output src/generated_contract_wrappers --backend ethers && prettier --write 'src/generated_contract_wrappers/**.ts'",
         "lint": "tslint --project .",
         "test:circleci": "run-s test:coverage",
         "test": "yarn run_mocha",
         "rebuild_and_test": "run-s build test",
         "test:coverage": "nyc npm run test --all && yarn coverage:report:lcov",
         "coverage:report:lcov": "nyc report --reporter=text-lcov > coverage/lcov.info",
-        "clean": "shx rm -rf _bundles lib test_temp scripts src/contract_wrappers/generated",
+        "clean": "shx rm -rf _bundles lib test_temp scripts src/generated_contract_wrappers",
         "build:umd:prod": "NODE_ENV=production webpack",
         "build:commonjs": "tsc && copyfiles -u 2 './src/compact_artifacts/**/*.json' ./lib/src/compact_artifacts && copyfiles -u 3 './lib/src/monorepo_scripts/**/*' ./scripts",
         "run_mocha": "mocha lib/test/**/*_test.js lib/test/global_hooks.js --timeout 10000 --bail --exit",
@@ -54,9 +54,9 @@
                     "../contract-wrappers/src/contract_wrappers/token_transfer_proxy_wrapper.ts",
                     "../contract-wrappers/src/contract_wrappers/token_wrapper.ts",
                     "../order-watcher/src/order_watcher/order_watcher.ts",
-                    "./src/contract_wrappers/generated/ether_token.ts",
-                    "./src/contract_wrappers/generated/token.ts",
-                    "./src/contract_wrappers/generated/exchange.ts"
+                    "./src/generated_contract_wrappers/ether_token.ts",
+                    "./src/generated_contract_wrappers/token.ts",
+                    "./src/generated_contract_wrappers/exchange.ts"
                 ],
                 "s3BucketPath": "s3://doc-jsons/0x.js/",
                 "s3StagingBucketPath": "s3://staging-doc-jsons/0x.js/"
