+++ packages/contracts/test/asset_proxy_owner.ts
@@ -12,8 +12,8 @@ import {
     ExecutionContractEventArgs,
     ExecutionFailureContractEventArgs,
     SubmissionContractEventArgs,
-} from '../src/contract_wrappers/generated/asset_proxy_owner';
-import { MixinAuthorizableContract } from '../src/contract_wrappers/generated/mixin_authorizable';
+} from '../src/generated_contract_wrappers/asset_proxy_owner';
+import { MixinAuthorizableContract } from '../src/generated_contract_wrappers/mixin_authorizable';
 import { artifacts } from '../src/utils/artifacts';
 import { chaiSetup } from '../src/utils/chai_setup';
 import { constants } from '../src/utils/constants';
