+++ packages/contracts/test/multi_sig_with_time_lock.ts
@@ -10,7 +10,7 @@ import * as Web3 from 'web3';
 import {
     MultiSigWalletWithTimeLockContract,
     SubmissionContractEventArgs,
-} from '../src/contract_wrappers/generated/multi_sig_wallet_with_time_lock';
+} from '../src/generated_contract_wrappers/multi_sig_wallet_with_time_lock';
 import { artifacts } from '../src/utils/artifacts';
 import { chaiSetup } from '../src/utils/chai_setup';
 import { constants } from '../src/utils/constants';
